Cleanup and adding all of the some commas
