+++ lib/account/account_page.dart
@@ -17,11 +17,13 @@ class _AccountPageState extends State<AccountPage> {
   @override
   Widget build(BuildContext context) {
     Widget profilePic = CircleAvatar(
-        backgroundColor: Theme.of(context).colorScheme.secondary,
-        child: Text('U'));
+      backgroundColor: Theme.of(context).colorScheme.secondary,
+      child: const Text('U'),
+    );
+
     Widget userName = Column(
       children: <Widget>[
-        Text("User"),
+        const Text("User"),
         TextButton.icon(
           onPressed: signOut,
           icon: const Icon(Icons.logout),
@@ -29,11 +31,14 @@ class _AccountPageState extends State<AccountPage> {
         )
       ],
     );
+
     Widget userRow = InkWell(
-        onTap: () => print("User Menu"),
-        child: Row(
-          children: <Widget>[profilePic, userName],
-        ));
+      onTap: () => {},
+      child: Row(
+        children: <Widget>[profilePic, userName],
+      ),
+    );
+
     return Scaffold(
       appBar: AppBar(
         title: const Text("Account"),
@@ -42,10 +47,14 @@ class _AccountPageState extends State<AccountPage> {
         children: [
           userRow,
           ListTile(
-              leading: const Icon(Icons.settings),
-              title: const Text('Settings'),
-              onTap: () => Navigator.of(context).push(MaterialPageRoute(
-                  builder: (context) => const SettingsPage()))),
+            leading: const Icon(Icons.settings),
+            title: const Text('Settings'),
+            onTap: () => Navigator.of(context).push(
+              MaterialPageRoute(
+                builder: (context) => const SettingsPage(),
+              ),
+            ),
+          ),
         ],
       ),
     );
