+++ lib/account/pages/settings_page.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/util/supabase.dart';
 
 class SettingsPage extends StatefulWidget {
   const SettingsPage({super.key});
