+++ lib/main_app.dart
@@ -3,7 +3,6 @@ import 'package:flutter_app/account/account_page.dart';
 import 'package:flutter_app/drives/pages/drives_page.dart';
 import 'package:flutter_app/home_page.dart';
 import 'package:flutter_app/rides/pages/rides_page.dart';
-import 'package:flutter_app/account/pages/settings_page.dart';
 
 enum TabItem { home, drives, rides, account }
 
@@ -41,52 +40,53 @@ class _MainAppState extends State<MainApp> {
   @override
   Widget build(BuildContext context) {
     return WillPopScope(
-        onWillPop: () async {
-          final isFirstRouteInCurrentTab =
-              !await _navigatorKeys[_currentTab]!.currentState!.maybePop();
-          if (isFirstRouteInCurrentTab) {
-            if (_currentTab == TabItem.home) {
-              return true;
-            }
-
-            _selectTab(TabItem.home);
+      onWillPop: () async {
+        final isFirstRouteInCurrentTab =
+            !await _navigatorKeys[_currentTab]!.currentState!.maybePop();
+        if (isFirstRouteInCurrentTab) {
+          if (_currentTab == TabItem.home) {
+            return true;
           }
-          return false;
-        },
-        child: Scaffold(
-          body: IndexedStack(
-            index: _currentTab.index,
-            children: TabItem.values
-                .map((tabItem) => buildNavigatorForTab(tabItem))
-                .toList(),
-          ),
-          bottomNavigationBar: BottomNavigationBar(
-            type: BottomNavigationBarType.fixed,
-            items: const <BottomNavigationBarItem>[
-              BottomNavigationBarItem(
-                icon: Icon(Icons.home),
-                label: 'Home',
-              ),
-              BottomNavigationBarItem(
-                icon: Icon(Icons.drive_eta),
-                label: 'Drives',
-              ),
-              BottomNavigationBarItem(
-                icon: Icon(Icons.chair),
-                label: 'Rides',
-              ),
-              BottomNavigationBarItem(
-                icon: Icon(Icons.account_circle),
-                label: 'Account',
-              ),
-            ],
-            currentIndex: _currentTab.index,
-            selectedItemColor: Colors.blue,
-            onTap: (index) {
-              _selectTab(TabItem.values[index]);
-            },
-          ),
-        ));
+
+          _selectTab(TabItem.home);
+        }
+        return false;
+      },
+      child: Scaffold(
+        body: IndexedStack(
+          index: _currentTab.index,
+          children: TabItem.values
+              .map((tabItem) => buildNavigatorForTab(tabItem))
+              .toList(),
+        ),
+        bottomNavigationBar: BottomNavigationBar(
+          type: BottomNavigationBarType.fixed,
+          items: const <BottomNavigationBarItem>[
+            BottomNavigationBarItem(
+              icon: Icon(Icons.home),
+              label: 'Home',
+            ),
+            BottomNavigationBarItem(
+              icon: Icon(Icons.drive_eta),
+              label: 'Drives',
+            ),
+            BottomNavigationBarItem(
+              icon: Icon(Icons.chair),
+              label: 'Rides',
+            ),
+            BottomNavigationBarItem(
+              icon: Icon(Icons.account_circle),
+              label: 'Account',
+            ),
+          ],
+          currentIndex: _currentTab.index,
+          selectedItemColor: Colors.blue,
+          onTap: (index) {
+            _selectTab(TabItem.values[index]);
+          },
+        ),
+      ),
+    );
   }
 
   Widget buildNavigatorForTab(TabItem tabItem) {
