+++ lib/util/big_button.dart
@@ -1,15 +1,14 @@
 import 'package:flutter/material.dart';
 
-class SubmitButton extends StatelessWidget {
+class BigButton extends StatelessWidget {
   final String text;
   final Function()? onPressed;
-
-  const SubmitButton({super.key, required this.text, this.onPressed});
+  const BigButton({super.key, required this.text, this.onPressed});
 
   @override
   Widget build(BuildContext context) {
     return SizedBox(
-      width: 170,
+      width: double.infinity,
       height: 53,
       child: MaterialButton(
         shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(100)),
@@ -18,8 +17,9 @@ class SubmitButton extends StatelessWidget {
         child: Text(
           text,
           style: TextStyle(
-              color: Theme.of(context).colorScheme.onPrimary,
-              fontWeight: FontWeight.w500),
+            color: Theme.of(context).colorScheme.onPrimary,
+            fontWeight: FontWeight.w500,
+          ),
         ),
       ),
     );
