+++ lib/util/loading_button.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/own_theme_fields.dart';
+import 'package:flutter_app/util/own_theme_fields.dart';
 import 'package:progress_state_button/iconed_button.dart';
 import 'package:progress_state_button/progress_button.dart';
 
@@ -29,31 +29,39 @@ class _LoadingButtonState extends State<LoadingButton> {
     return ProgressButton.icon(
       iconedButtons: {
         ButtonState.idle: IconedButton(
-            text: widget.idleText ?? "Submit",
-            icon: widget.idleIcon ??
-                Icon(Icons.send,
-                    color: Theme.of(context).colorScheme.onPrimary),
-            color: Theme.of(context).colorScheme.primary),
+          text: widget.idleText ?? "Submit",
+          icon: widget.idleIcon ??
+              Icon(
+                Icons.send,
+                color: Theme.of(context).colorScheme.onPrimary,
+              ),
+          color: Theme.of(context).colorScheme.primary,
+        ),
         ButtonState.loading:
             IconedButton(color: Theme.of(context).colorScheme.primary),
         ButtonState.fail: IconedButton(
-            text: widget.failText ?? "Failed",
-            icon: Icon(Icons.cancel,
-                color: Theme.of(context).colorScheme.onError),
-            color: Theme.of(context).colorScheme.error),
+          text: widget.failText ?? "Failed",
+          icon: Icon(
+            Icons.cancel,
+            color: Theme.of(context).colorScheme.onError,
+          ),
+          color: Theme.of(context).colorScheme.error,
+        ),
         ButtonState.success: IconedButton(
-            text: widget.successText ?? "Success",
-            icon: Icon(
-              Icons.check_circle,
-              color: Theme.of(context).own().onSuccess,
-            ),
-            color: Theme.of(context).own().success)
+          text: widget.successText ?? "Success",
+          icon: Icon(
+            Icons.check_circle,
+            color: Theme.of(context).own().onSuccess,
+          ),
+          color: Theme.of(context).own().success,
+        )
       },
       onPressed: widget.onPressed,
       state: widget.state,
       textStyle: TextStyle(
-          color: Theme.of(context).colorScheme.onPrimary,
-          fontWeight: FontWeight.w500),
+        color: Theme.of(context).colorScheme.onPrimary,
+        fontWeight: FontWeight.w500,
+      ),
     );
   }
 }
